requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
bcrypt==4.2.1
click==8.1.7
fastapi==0.115.5
h11==0.14.0
idna==3.10
Jinja2==3.1.4
jose==1.0.0
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
python-multipart==0.0.19
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
